Rename ConfigurationManager to JsonHelper
